# data_aggregator_service/data-aggregator-service-openapi.yml
openapi: 3.0.3
info:
  title: Data Aggregator Service API
  description: |
    Микросервис для агрегации данных о перемещениях объектов, генерации отчетов и аналитики.
    Периодически собирает данные от positioning-service, агрегирует их и предоставляет аналитические отчеты.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: analytics@rtls.example.com
servers:
  - url: http://localhost:8085/api/v1
    description: Локальный сервер разработки
  - url: https://analytics.rtls.example.com/api/v1
    description: Продуктивный сервер
...
paths:
...
  /analytics/anomalies:
    get:
      tags:
        - Analytics
      summary: Обнаружение аномалий в поведении
      description: Выявление аномалий в перемещениях сотрудников и оборудовании
      operationId: detectAnomalies
      parameters:
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: entity_ids
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: anomaly_types
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
              enum: [unexpected_zone, unusual_time, abnormal_speed, prolonged_stay]
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyDetectionReport'
        '400':
          description: Некорректные параметры запроса
...
components:
  schemas:
...
    AnomalyDetectionReport:
      type: object
      properties:
        report_id:
          type: string
        generated_at:
          type: string
          format: date-time
        period:
          type: object
          properties:
            start_time:
              type: string
              format: date-time
            end_time:
              type: string
              format: date-time
        anomalies:
          type: array
          items:
            type: object
            properties:
              anomaly_id:
                type: string
              timestamp:
                type: string
                format: date-time
              anomaly_type:
                type: string
                enum: [unexpected_zone, unusual_time, abnormal_speed, prolonged_stay]
              entity_id:
                type: string
              entity_name:
                type: string
              entity_type:
                type: string
              zone_id:
                type: string
                nullable: true
              zone_name:
                type: string
                nullable: true
              position:
                type: object
                properties:
                  x:
                    type: number
                  y:
                    type: number
                  z:
                    type: number
              description:
                type: string
              severity:
                type: string
                enum: [low, medium, high, critical]
              confidence:
                type: number
                minimum: 0
                maximum: 1
              related_violations:
                type: array
                items:
                  type: string
...